Replaced keys by actions
also: decalage is now CaseWidget.tShift everywhere.
fixed the display of "Quality" in GUI
added method generate_authors (which looks at self.savefolder and add
the authors there)
One need's to call CaseAnalyserWidget.generate_authors_cases() to load
the author cases.
chg_savefolder was added to CaseAnalyserWidget (removing the GUI action
on labels). It calls generate_authors and generate_authors_cases.

+++ kg/algorithm.py
@@ -281,7 +281,7 @@ class ZischenDetetkt2(Algorithm):
     def __init__(self, fc, threshold, dt, Rexport=False):
         #
         param = {'fmin': 100,'fmax': 15000, 'overlap': 6} 
-        param['threshold'] = int(threshold)
+        param['threshold'] = float(threshold)
         param['fc']= int(fc)
         param['dt']= float(dt)
         description = """implement the Algorithm:
